Improve benchmarks
Add the comprehensive benchmark from perf/1-1.  This is intended to be a point
of comparison to ensure we don't have any major regressions when switching to
the token API.

Also fix the constructor benchmarks.  Initially bench/node.js showed a massive
regression in perf/1-1, but irhydra2 demonstrated that this is because the
original benchmark was a) inlining the constructor and b) not actually doing any
writes.

Finally add a few more benchmarks similar to comprehensive but with stats use.